requirements.txt
tensorflow
numpy
matplotlib
pillow
scikit-learn
opencv-python
